requirements.txt
aiogram==3.13.1
pydantic-settings==2.6.1
openai>=1.60.0
aiosqlite==0.20.0
SQLAlchemy==2.0.36
APScheduler==3.10.4
anthropic==0.40.0
httpx==0.27.2
python-dotenv==1.0.1
aiohttp==3.10.11
